# internal/api/go/model_login_user_request.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * User Service API
 *
 * API of User microservice
 *
 * API version: 0.0.1
 */

package userApi




type LoginUserRequest struct {

	Email string `json:"email"`

	Password string `json:"password"`
}

// AssertLoginUserRequestRequired checks if the required fields are not zero-ed
func AssertLoginUserRequestRequired(obj LoginUserRequest) error {
	elements := map[string]interface{}{
		"email": obj.Email,
		"password": obj.Password,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertLoginUserRequestConstraints checks if the values respects the defined constraints
func AssertLoginUserRequestConstraints(obj LoginUserRequest) error {
	return nil
}